Include test cases to existing classes
