+++ tests/jballerina-unit-test/src/test/java/org/ballerinalang/test/query/SimpleQueryExpressionWithDefinedTypeTest.java
@@ -269,5 +269,83 @@ public class SimpleQueryExpressionWithDefinedTypeTest {
         Assert.assertEquals(returnValues.length, 1, "Expected events are not received");
         Assert.assertTrue(returnValues[0] instanceof BError, "Expected BErrorType type value");
     }
-}
+	
+	@Test(description = "Test anonymous record type, record type referencing, optional field, " +
+            "changed order of the fields")
+    public void testOthersAssociatedWithRecordTypes() {
+        BValue[] returnValues = BRunUtil.invoke(result, "testOthersAssociatedWithRecordTypes");
+        Assert.assertNotNull(returnValues);
 
+        Assert.assertEquals(returnValues.length, 2, "Expected events are not received");
+
+        BMap teacher1 = (BMap) returnValues[0];
+        BMap teacher2 = (BMap) returnValues[1];
+
+        Assert.assertTrue(teacher1.get("classStudents") instanceof BValueArray);
+        Assert.assertTrue(teacher1.get("experience") instanceof BMap);
+        Assert.assertEquals(teacher1.get("classStudents").stringValue(),
+                "[{firstName:\"Alex\", lastName:\"George\", score:82.5}, " +
+                        "{firstName:\"Ranjan\", lastName:\"Fonseka\", score:90.6}]");
+        Assert.assertEquals(teacher1.get("experience").stringValue(), "{duration:10, qualitifications:\"B.Sc.\"}");
+        Assert.assertEquals(teacher1.stringValue(),
+                "{classStudents:[{firstName:\"Alex\", lastName:\"George\", score:82.5}, " +
+                        "{firstName:\"Ranjan\", lastName:\"Fonseka\", score:90.6}], " +
+                        "experience:{duration:10, qualitifications:\"B.Sc.\"}, " +
+                        "firstName:\"Alex\", lastName:\"George\", deptAccess:\"XYZ\", " +
+                        "address:{city:\"NY\", country:\"America\"}}");
+
+        Assert.assertTrue(teacher2.get("classStudents") instanceof BValueArray);
+        Assert.assertTrue(teacher2.get("experience") instanceof BMap);
+        Assert.assertEquals(teacher2.get("classStudents").stringValue(),
+                "[{firstName:\"Alex\", lastName:\"George\", score:82.5}, " +
+                        "{firstName:\"Ranjan\", lastName:\"Fonseka\", score:90.6}]");
+        Assert.assertEquals(teacher2.get("experience").stringValue(),
+                "{duration:10, qualitifications:\"B.Sc.\"}");
+        Assert.assertEquals(teacher2.stringValue(),
+                "{classStudents:[{firstName:\"Alex\", lastName:\"George\", score:82.5}, " +
+                        "{firstName:\"Ranjan\", lastName:\"Fonseka\", score:90.6}], " +
+                        "experience:{duration:10, qualitifications:\"B.Sc.\"}, " +
+                        "firstName:\"Ranjan\", lastName:\"Fonseka\", deptAccess:\"XYZ\", " +
+                        "address:{city:\"NY\", country:\"America\"}}");
+    }
+
+    @Test(description = "Test query expressions with tuple typed binding in let")
+    public void testQueryExprTupleTypedBinding2() {
+        BValue[] returnValues = BRunUtil.invoke(result, "testQueryExprTupleTypedBinding2");
+        Assert.assertNotNull(returnValues);
+
+        Assert.assertEquals(returnValues.length, 1, "Expected events are not received");
+        Assert.assertEquals(((BValueArray) returnValues[0]).getInt(0), 3);
+    }
+
+    @Test(description = "Test query expression with type conversion in select clause")
+    public void testQueryExprWithTypeConversion() {
+        BValue[] returnValues = BRunUtil.invoke(result, "testQueryExprWithTypeConversion");
+        Assert.assertNotNull(returnValues);
+
+        Assert.assertEquals(returnValues.length, 2, "Expected events are not received");
+
+        BMap person1 = (BMap) returnValues[0];
+        BMap person2 = (BMap) returnValues[1];
+
+        Assert.assertEquals(person1.stringValue(), "{firstName:\"Alex\", lastName:\"George\", " +
+                "deptAccess:\"XYZ\", " +
+                "address:{city:\"New York\", country:\"America\"}}");
+        Assert.assertEquals(person2.stringValue(), "{firstName:\"Ranjan\", lastName:\"Fonseka\", " +
+                "deptAccess:\"XYZ\", " +
+                "address:{city:\"New York\", country:\"America\"}}");
+    }
+
+    @Test(description = "Test streams with map and filter")
+    public void testQueryExprWithStreamMapAndFilter() {
+        BValue[] returnValues = BRunUtil.invoke(result, "testQueryExprWithStreamMapAndFilter");
+        Assert.assertNotNull(returnValues);
+
+        Assert.assertEquals(returnValues.length, 1, "Expected events are not received");
+
+        BMap subscription = (BMap) returnValues[0];
+
+        Assert.assertEquals(subscription.stringValue(), "{firstName:\"Ranjan\", lastName:\"Fonseka\", score:90.6, " +
+                "degree:\"Bachelor of Medicine\"}");
+    }
+}
